requirements.txt
python-dotenv~=1.0.1
Flask~=3.1.0
mysql~=0.0.3
mysql-connector-python~=9.1.0
SQLAlchemy~=2.0.36
pymysql
alembic~=1.14.0
cryptography~=44.0.0
bcrypt~=4.2.1
mysqlclient~=2.2.6
jwcrypto~=1.5.6
Werkzeug~=3.1.3
requests~=2.32.3
PyJWT==1.7.1
jwt==1.3.1
Flask-Cors~=5.0.0